Added very hackish command parsing to shell
The applications don't have a local heap yet, so it's not possible
to produce a Vec of words from a given string.  I only need the
command name for now (arguments are not yet supported by exec())
so I just separate the first word and copy it into another buffer
with the path added at the start (always "/bin/").  I can now
actually specify a binary on the command line to run