Added command line argument passing to elf loader
It will now take the command line arguments and enviroment variables
as an argument to the exec() system call, which will encode them as
C ABI-style argv and envp values at the top of the stack, and then
pass them into the newly-executed binary via the CPU registers, by
writing to the Context before returning.  In order to avoid the
syscall return overwriting the context return values, the exec()
syscall is created as different, and doesn't return as the rest of
the syscalls do.

It's clumsy because manipulating null terminated strings in Rust is
not trivial, and the ffi::Cstr struct is not available.  I chose to
use the C ABI format of passing arguments to main because I'm using
the C/Unix ABI syscalls, even though the ruxpin_app crate wraps them
in Args and Vars and doesn't pass them directly to main().  It made
sense to keep it the same, even though it was more work.  It also
sidesteps the issue of how fat pointers are represented in Rust as
passed through registers to a function (eg. is the pointer in x0,
size in x1, and then the second argument starts in x2?)  It works
for now at least