Fixed the spinlock timeout
This fixes the spinlock timeout by removing the "open" call from
the process files table, so that the process lock is not held for the
entirety of the open, when the filename data is accessed from user
space.  Now, it can find an open slot, and save it separately, and
use the vfs::open call to do the actual open.

I was also thinking of moving the spinlock on each process down to
its fields, so that the ".files" and ".space" fields could be locked
separately.  That would be safer overall (prevent a timeout) but it
means more locks, and I need to settle a few more things in Process
before I don that.

I also added ".try_lock().unwrap()" to a bunch of things so that the
panic occurs where the lock is tried, and not in the lock code, which
isn't helpful.  I'm only getting the location of the panic from the
rust panic handler and not the full stack trace.