Added start of canonical input processing in tty
I added the tasklets system in order to process the key after an
interrupt occurs, but it's not working when I use it, so the key
processing is happening during the irq handler for the console
driver.  It will read in a line and then wait until the process
reads the line before accepting more input, which works for now.

I also added a ProcessState enum to prevent accidentally removing
and readding a process from the wrong list, and change the suspend
call to take a process instead of using the current, for the case
of being used in a tasklet where the current process might have
changed by then.