Fixed issue with processing char as a tasklet
It turned out I was scheduling a tasklet inside of a tasklet, and
for some reason that didn't work, even though the tasklets is using
`.pop_back()` (now `.pop_front()`) to get the element instead of an
iterator or something that would explain the scheduled tasklet not
running.  I had actually solved the problem before I realized I did
due to the fact that printing during the serial interrupt locks up,
and I happened to have a print debug statement in the restart func.

I also moved the schedule function call to the timer interrupt, so
it wont happen when the serial interrupt occurs