Replaced CircularBuffer with VecDequeue
Since the constructor for VecDequeue isn't const, I had to rewrite the
console driver a bit so that it doesn't need the buffer until normal
mode (when the allocator is available).  Since the writing functions
don't use `self` and instead use a static DeviceRegisters, this was
a fairly easy change, once I realized it wasn't printing anything
because the allocator wasn't available before the printing starts.