Added sbrk and application heap
It's very hacky for the moment, but it works enough that I can use
the format macro to build a filename which is then opened.  I added
a ps command which uses this (although it's not entirely necessary
to use the heap there).  The actual heap implementation a copy of
kmalloc, which itself is a direct rust port of the C allocator from
ComputieOS, so it should probably be replaced eventually.  It
calls the sbrk syscall to increase the data segment size, which is
determined to be the highest program segment that isn't the stack.
It doesn't adjust the stack segment, which is only 4k*4k anyways,
so it could crash into the stack unchecked if enough memory is
requested.  This is only meant as a stop gap so that I don't have
to write all the applications without String and Vec, which are
really useful to have.