Major reorganization of files into crates
I've been meaning to do this for a while, but it's finally time.
This change will break the kernel up, such that the kernel will now
be a library that's consumed by a specific config/ crate, which is
tailored to a specific board/platform.  It also moves the implementation
of drivers and filesystems out of the kernel, so that they can more
easily be selectively compiled in for a given platform.