Refactored mmu translation table code
In order to reuse the translation table traversal code, I modified
it to use a visitor with a trait and special type defined for each
function that wants to traverse the table.  The unmap code now uses
this, as does a new function to modify permissions in a range, but
the map code doesn't use it (at least not yet) because the traversal
is a bit different since it wants to hit empty table areas, not
occupied ones like every other traversal.

I also threw in a fix for system calls such that if the process
changes during a syscall, the result from that syscall will be
written to the process that started the syscall and not the one
that was switched to.  Most syscalls return to the same process
so it didn't cause a problem much, but to test another problem,
I needed to be able to switch to the child process after the fork.