Reintegrated the bitmap code into pages
I can factored it out into a misc/ module, so I could reuse it for
the inodes, but that requires a more filesystem specific implementation
that didn't fit well to the generic version, so the paging system was
the only place it was used.  Since I wanted to add page descriptors
as well, so that I can have a per-page refcount, it was easier to put
it all in the same file again.