Modified to use page refcount for freeing
The reliance on Arc<Segment> has been removed.  I haven't yet actually
removed Arc<Segment> since it might still be used to keep track of that
segment's page cache or page map, which would be shared between processes
but with the freeing of pages tied to the page and not the segment, it's
now possible to implement copy on write.

One thing not fixed by this is sharing pages that haven't yet been
loaded.  There is no mechanism to map a new page into multiple
existing segments since there is no tracking of which processes
share a segment, but this will eventually be replaced by a page
cache for each open file, such that when a page is accessed that isn't
yet loaded, it will first check the cache and load the page if found
(incrementing the refcount) or else load the page into the cache so
that any other process that loads the same page will share the same
page in memory as well (for read only segments)