Fixed issue with adjacent pages being zeroed
This is the problem that's been plaguing me for the last week and a
half.  I was getting a very strange error, aarch64 exception 0b000000.
It's documented as "Unknown Exception", and has no information on what
causes it because the cause is unknown!  Qemu was reporting it as an
Undefined Instruction, and it took a while before I modified the fatal
error printer to print the instruction at the ELR address, which showed
0.  But everything indicated it should have been loaded.

It turns out it was a pointer error... why am I not surprised.  It was
caused by code that creates a slice that is then zeroed when allocating
pages.  The pointer is assumed to be i32 because it's not specified I
guess, and that means it overwrites 4 adjacent pages.  So when it was
loading address 0x1051000, it was overwriting 0x1053000, which had the
executable code for the shell program. *sigh*