Removed arch-specific idle process
Previously it was directly injecting instructions for aarch64 into
memory and executing that as the idle process. It now creates a
kernel process at startup, which uses the kernel translation tables
and can be created with a function pointer as the address to start
running at.  In future, it will be possible to make other threads
that inherit from the same process.

It currently executes unpriviledged because setting the SPSR reg
to either of the EL1 modes caused different problems (one panicked
and the other seemed to loop without executing the shell at all