Added log levels to control kernel output
There are a number of log levels, not all of which are used yet.
Error, Warning, Notice, Info, Debug, Trace.  Notice is for normal
messages that should always be printed.  Info, Debug, and Trace
are for debugging messages in increasing level of verbosity.  Trace
is meant to only be used as a last resort, whereas debug should
still be usable as such.  Info is for detailed but not debugging-
specific messages.