Added sync syscall and hooked up fs writing
It still doesn't work properly, but it sort of works.  Writing of
directories seems to work, inodes are created, but contents of
directories and files doesn't seem to be present.  It's now possible
to test though, which is the important part.  I've also made it
possible to boot without reinitializing the disk.