Added hacky rename() for ext2
There are 3 vnodes involved in a rename: the parent directory of the
existing file, the parent directory of the destination new file, and
the vnode of the file being renamed.  For ext2, only the inode of
the source file is needed, and if the two parents are the same vnode,
it's fairly easy to make work, but if they are different, you can't
get both unwrapped at the same time.  You can only get a concrete
Ext2Vnode when inside a VnodeOperation function called on one of the
vnodes.

To get around this, I'm using the fact that hard linking on ext2
is available, which uses the same disk inode pointed to by two file
entries.  By linking to the new location of the file, and then
unlinking from the original location, I can transfer the inode from
one to the other, using only public functions.