Fixed issue with incorrect return from blocking
The shell was printing the prompt before after pressing enter, and
then running the command and printing the prompt again after.  It
was because the read syscall was returning 0 length after the enter
key had been pressed, which was because it didn't restart the syscall.
The syscall would only be restarted after a schedule() and not after
an irq, and it was during the irq handling that the shell process
would be unblocked, so it would return 0 the first time, the shell
would loop back and then read again, which would complete successfully.
It now checks for a restart at the end of any user exception or irq.

I also threw some fixes in for warnings in ext2, and some erroneous
tabs mixed in with the spaces