Added . and .. support for most file systems
It might still be a bit buggy, but all vnodes now have a `.set_self()`
method that takes a weak pointer to the vnode itself, which can be
saved and used for . and .. for devfs, tmpfs, and procfs.  It can
also be used in `.create()` to assign the parent to the newly created
vnode.  When mounting a file system, the parent of the mount point,
rather than the mount point itself, is passed in, and can be used to
directly link to the mount point's parent directory.  For ext2, there
is a special exception that uses the `mounted_on` field to get the
parent vnode, but the other systems have had that field removed,
since it's no longer needed