Attempted to boot the secondary cores
I'm getting that same error from before, where ESR = 0x02000000 or
0b000000, which is unknown error.  The actual error occurs when the
secondary core opens a file or prints a message, but the fatal kernel
error is during the schedule() function run on the boot core, cpu 0.
Not yet sure what the problem is, but I wanted to commit my work in
progress.  The secondary cores aren't being run, unless the code in
start.s is uncommented.