Added page cache
When a file is executed, the page will be entered into the page cache,
indexed using the vnode's Arc address.  Pages can be requested from
the cache, which will either be returned if already loaded, or loaded
from the file into a new page and entered into the cache list.
When a page miss occurs, it will try to fetch the page from the cache
instead of loading it, which means all processes that use the same
executable file will share the pages of that file.  The pages are
set to copy on write if the segment is ReadWrite, so that the cached
pages are always identical to the file on disk.  This fixes the
issue where pages loaded from a file after a process forks would not
be shared between the processes.

There is currently no way to unload a cached file, so technically,
it's leaking memory atm.