Fixed sbrk to adjust both stack and data segments
The segments are now ordered by mapping, and since the stack is
normally placed at the upper limit of memory, the stack should be
the last segments.  If it's a stack segment and the previous segment
is a data segment, then it adjusts the data segment, and if it
overlaps with the stack segment, it adjusts the stack segment too.

There is still no stack page guard.