Refactored mmu table mapping to use the visitor
The previous versions were actually pretty broken, which is why they
were so odd.  The non-mapping visitor was refactored to the point
where it was pretty easy to understand, and then a mapping visitor
was created from that which can do the same as the old mapping code,
but is now a bit more controllable.

The end goal here is to allocate tables on demand, as well as the
blocks in those tables