Refactored the copy memory mapping functions
Hopefully this is the last refactor before I can make the tables
allocated on demand as well as the pages.  I created a new visitor
for two tables, so that the recursive arguments can be used to keep
track of the table references.  The actual trait implementation is
much simpler and easier to reason about as a result.  Using the
existing visitor for this required a bunch of weird exceptions and
I didn't get it fully working before abandoning it.