Modified to allocate translation tables on demand
Previously it was allocating tables for the entire mapped region
of a segment, leaving only the 4KB pages unallocated until accessed,
but for very large ranges, that's still a lot of pages.  This change
now only allocates tables if any of the sub addresses of that table
are different.  If they are all the same (eg. all 4KB blocks in the
same segment with the same permissions, etc), then it doesn't
allocate the table, but does set the flags (similar to how a block
entry has an address of 0, no access flag, but the permissions
flags are set).  When updating an address due to an access fault, it
will allocate the table and copy the permissions from the parent
descriptor into all entries in the table.

It should now be possible to allocate a stack segment containing
all of memory